+++ docs/src/surfaces/actionBar/actionBar.jsx
@@ -2,6 +2,10 @@ import {
     TitleBar,
     Typography,
 } from 'react-cm-ui';
+import {
+    camelCase,
+    startCase,
+} from 'lodash';
 import React from 'react';
 import ComponentApi from '../../global/componentApi';
 import Heading from '../../global/heading';
@@ -12,11 +16,14 @@ import { default as actionBarDoc } from '!!@advclb/react-docgen-loader!react-cm-
 /* eslint-enable import/no-named-default, import/extensions */
 
 function DocsActionBar() {
-    const descriptionCopy = actionBarDoc.description;
+    const {
+        description,
+        displayName,
+    } = actionBarDoc;
 
     return (
-        <Main page="action_bar">
-            <TitleBar title="Action Bar" />
+        <Main page={camelCase(displayName)}>
+            <TitleBar title={startCase(displayName)} />
 
             <Main.Content>
                 <MarkdownContainer>
@@ -24,7 +31,7 @@ function DocsActionBar() {
                         className="description"
                         variant="body1"
                     >
-                        {descriptionCopy}
+                        {description}
                     </Typography>
 
                     <Heading