+++ src/global/constants.js
@@ -2,15 +2,21 @@ export const ENTER_KEY_CODE = 13;
 
 export const UI_CLASS_NAME = 'cmui';
 
+export const BEM_DATA_GRID = 'data_grid';
+
 export const BEM_EMAIL_LINK = 'email_link';
+export const BEM_NAVIGATION_BREADCRUMBS = 'navigation_breadcrumbs';
+export const BEM_NAVIGATION_SECTIONAL_TABS = 'navigation_sectional_tabs';
+export const BEM_NAVIGATION_TAB_ROOT_CLASS = `${BEM_NAVIGATION_SECTIONAL_TABS}--tab`;
 export const BEM_PERSON_CONTACT_INFO = 'person_contact_info';
-export const BEM_TELEPHONE_LINK = 'telphone_link';
 export const BEM_PERSON_CORE_MILESTONES = 'person_core_milestones';
 export const BEM_PERSON_PANEL = 'person_panel';
 export const BEM_PERSON_PANEL_DETAILS = `${BEM_PERSON_PANEL}--details`;
 export const BEM_PERSON_PANEL_DETAILS_ACTION_BUTTON = `${BEM_PERSON_PANEL_DETAILS}_action_button`;
+export const BEM_TELEPHONE_LINK = 'telphone_link';
 
 export const EMAIL_LINK_CLASSES = `${UI_CLASS_NAME} ${BEM_EMAIL_LINK}`;
 export const PERSON_CORE_MILESTONES_CLASSES = `${UI_CLASS_NAME} ${BEM_PERSON_CORE_MILESTONES}`;
 export const PERSON_CONTACT_INFO_CLASSES = `${UI_CLASS_NAME} ${BEM_PERSON_CONTACT_INFO}`;
 export const TELEPHONE_LINK_CLASSES = `${UI_CLASS_NAME} ${BEM_TELEPHONE_LINK}`;
+