+++ src/inputs/dropdownButton/dropdownButton.jsx
@@ -33,12 +33,25 @@ const propTypes = {
      * The ButtonDropdown's label.
      */
     label: PropTypes.string,
+    /**
+     * Options list theme
+     */
+    optionsTheme: PropTypes.oneOf([
+        'dark',
+        'light',
+    ]),
+    /**
+     * Indicates whether or not the Dropdown Button can be focused.
+     */
+    tabIndex: PropTypes.number,
 };
 
 const defaultProps = {
     ...buttonDefaultProps,
     iconSize: 16,
     iconType: 'chevron-down',
+    optionsTheme: 'dark',
+    tabIndex: -1,
 };
 
 const useStyles = makeStyles({
@@ -67,8 +80,10 @@ function DropdownButton(props) {
         inverse,
         label,
         outlined,
+        optionsTheme,
         relax,
         style,
+        tabIndex,
         target,
         title,
         width,
@@ -84,8 +99,9 @@ function DropdownButton(props) {
 
     const bemClassName = 'button_dropdown';
     const rootClasses = ClassNames(
-        'button_dropdown',
+        bemClassName,
         className,
+        { [`${bemClassName}-open`]: isMenuOpen },
     );
 
     return (
@@ -110,6 +126,7 @@ function DropdownButton(props) {
             target={target}
             title={title}
             width={width}
+            tabIndex={tabIndex}
         >
             {iconType && (
                 <Icon
@@ -128,6 +145,7 @@ function DropdownButton(props) {
                 isOpen={isMenuOpen}
                 onToggleOpen={onMenuToggle}
                 getParentContainer={getParentContainer}
+                optionsTheme={optionsTheme}
             >
                 {children}
             </DropdownMenu>