+++ src/versions.js
@@ -1,7 +1,27 @@
 const versions = {
     'react-cm-ui': {
-        package: '9.15.2',
-        components: {},
+        package: '9.16.0',
+        components: {
+            inputs: {
+                durationPicker: {
+                    devLibraryVersion: '2.0.0',
+                    designLibraryVersion: '', // TODO: Add Design Library version
+                    designLibraryDoc: '', // TODO: Add Sketch Cloud Link
+                },
+            },
+            navigation: {
+                breadcrumbs: {
+                    devLibraryVersion: '2.0.0',
+                    designLibraryVersion: '', // TODO: Add Design Library version
+                    designLibraryDoc: '', // TODO: Add Sketch Cloud Link
+                },
+                sectionalTabs: {
+                    devLibraryVersion: '2.0.0',
+                    designLibraryVersion: '', // TODO: Add Design Library version
+                    designLibraryDoc: '', // TODO: Add Sketch Cloud Link
+                },
+            },
+        },
         theme: {
             breakpoints: '1.0.0',
             gutters: '1.0.0',